% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_answer}
\alias{export_answer}
\title{Export answers.}
\usage{
export_answer(x, file, format = "yaml")
}
\arguments{
\item{x}{object of class \code{form_answers}.}

\item{file}{file name.}

\item{format}{So far only YAML is available.}
}
\description{
Export answers.
}
