% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_among.R
\name{fob_among}
\alias{fob_among}
\alias{fob_month}
\alias{fob_weekday}
\title{A binary question to be answered by Yes or No.}
\usage{
fob_among(question, choices, field_name = "", output = "choice", ...)

fob_month(question, field_name = "", output = c("choice", "id"), ...)

fob_weekday(question, field_name = "", output = c("choice", "id"), ...)
}
\arguments{
\item{question}{The question to be asked.}

\item{choices}{A vector of choices.}

\item{field_name}{Name of the field (optional).}

\item{output}{Either "choice" or "id".}

\item{...}{further arguments to be passed to \code{\link[=generate_form_choices]{generate_form_choices()}}.}
}
\value{
Return the choice selected either as the choice value or its identifier (an integer).
}
\description{
A binary question to be answered by Yes or No.
}
\section{Functions}{
\itemize{
\item \code{fob_month}: One month.

\item \code{fob_weekday}: One weekday.
}}

