% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_any.R
\name{fob_any}
\alias{fob_any}
\title{Generate any kind of question.}
\usage{
fob_any(type, ...)
}
\arguments{
\item{type}{Type of answer expected.}

\item{...}{Further arguments passed to the question function.}
}
\description{
Generate any kind of question.
}
