% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_int.R
\name{fob_int}
\alias{fob_int}
\alias{fob_num}
\alias{fob_date}
\alias{fob_char}
\alias{fob_yorn}
\alias{fob_bool}
\title{Generate one question.}
\usage{
fob_int(question, field_name = "", prompt = NULL)

fob_num(question, field_name = "", prompt = NULL)

fob_date(question, field_name = "", format = "\%Y-\%m-\%d", prompt = NULL, ...)

fob_char(question, field_name = "", prompt = NULL)

fob_yorn(question, field_name = "", prompt = "Enter [Y]es or [N]o:", ...)

fob_bool(question, field_name = "", prompt = "Enter [T]RUE or [F]ALSE:", ...)
}
\arguments{
\item{question}{The question to be asked.}

\item{field_name}{Name of the field (optional).}

\item{prompt}{The string printed when prompting the user for input (see \code{\link[=readline]{readline()}}).}

\item{format}{Date format.}

\item{...}{Further arguments to be passed to \code{\link[=generate_form_pattern]{generate_form_pattern()}}.}
}
\value{
one integer.
}
\description{
Generate one question.
}
\section{Functions}{
\itemize{
\item \code{fob_num}: one numeric.

\item \code{fob_date}: one date.

\item \code{fob_char}: a character.

\item \code{fob_yorn}: \code{TRUE} for yes and \code{FALSE} for no.

\item \code{fob_bool}: one boolean/logical (i.e. \code{TRUE} or \code{FALSE}).
}}

