% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_pattern.R
\name{fob_pattern}
\alias{fob_pattern}
\title{A binary question to be answered by Yes or No.}
\usage{
fob_pattern(
  question,
  pattern,
  field_name = "",
  prompt = "Enter your answer:",
  ...
)
}
\arguments{
\item{question}{The question to be asked.}

\item{pattern}{Pattern to be matched.}

\item{field_name}{Name of the field (optional).}

\item{prompt}{the string printed when prompting the user for input (see \code{\link[=readline]{readline()}}).}

\item{...}{further arguments passed to \code{\link[=generate_form_pattern]{generate_form_pattern()}}.}
}
\value{
answer validated.
}
\description{
A binary question to be answered by Yes or No.
}
