% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_section.R
\name{fob_section}
\alias{fob_section}
\title{Add a section.}
\usage{
fob_section(msg)
}
\arguments{
\item{msg}{a character string that will be displayed.}
}
\value{
A \code{section} object.
}
\description{
Add a section.
}
