% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_word.R
\name{fob_word}
\alias{fob_word}
\alias{fob_text}
\alias{fob_email}
\title{A binary question to be answered by Yes or No.}
\usage{
fob_word(question, field_name = "", prompt = NULL, ...)

fob_text(question, field_name = "", max = NULL, prompt = NULL, ...)

fob_email(question, field_name = "", prompt = NULL, ...)
}
\arguments{
\item{question}{The question to be asked.}

\item{field_name}{Name of the field (optional).}

\item{prompt}{the string printed when prompting the user for input (see \code{\link[=readline]{readline()}}).}

\item{...}{further arguments to be passed to \code{\link[=generate_form_pattern]{generate_form_pattern()}}.}

\item{max}{maximum of characters (ignored if \code{NULL}).}
}
\value{
a word.
}
\description{
A binary question to be answered by Yes or No.
}
\section{Functions}{
\itemize{
\item \code{fob_text}: a character string.

\item \code{fob_email}: a character string.
}}

