% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pattern.R
\name{match_pattern}
\alias{match_pattern}
\title{A binary question to be answered by Yes or No.}
\usage{
match_pattern(question, pattern, post = NULL, field_name = "")
}
\arguments{
\item{question}{The question to be asked.}

\item{pattern}{Pattern to be matched.}

\item{post}{Post processing.}

\item{field_name}{Name of the field (optional).}
}
\value{
answer validated.
}
\description{
A binary question to be answered by Yes or No.
}
