% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\title{Repeat a form.}
\usage{
multi(x, n = 1, names = NULL)
}
\arguments{
\item{x}{a object of class \code{form}.}

\item{n}{number of participants.}

\item{names}{participants names.}
}
\description{
Fill out the same form by multiple participants.
}
