% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_among.R
\name{one_among}
\alias{one_among}
\title{A binary question to be answered by Yes or No.}
\usage{
one_among(question, choices, field_name = "", output = "choice", ...)
}
\arguments{
\item{question}{The question to be asked.}

\item{choices}{A vector of choices.}

\item{field_name}{Name of the field (optional).}

\item{output}{Either "choice" or "id".}

\item{...}{further arguments passed to generate_form_choices().}
}
\value{
Return the choice selected either as the choice value or its identifier (an integer).
}
\description{
A binary question to be answered by Yes or No.
}
