% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_integer.R
\name{one_integer}
\alias{one_integer}
\alias{one_numeric}
\alias{one_word}
\alias{one_text}
\alias{one_date}
\title{A binary question to be answered by Yes or No.}
\usage{
one_integer(question, field_name = "", prompt = NULL)

one_numeric(question, field_name = "", prompt = NULL)

one_word(question, field_name = "", prompt = NULL)

one_text(question, field_name = "", max = NULL, prompt = NULL)

one_date(question, field_name = "", format = "\%Y-\%m-\%d", prompt = NULL)
}
\arguments{
\item{question}{The question to be asked.}

\item{field_name}{Name of the field (optional).}

\item{prompt}{the string printed when prompting the user for input (see \code{\link[=readline]{readline()}}).}

\item{max}{maximum of characters (ignored if \code{NULL}).}

\item{format}{date format.}
}
\value{
Return the choice selected either as the choice value or its identifier (an integer).
}
\description{
A binary question to be answered by Yes or No.
}
\section{Functions}{
\itemize{
\item \code{one_numeric}: a numeric.

\item \code{one_word}: a word.

\item \code{one_text}: a text.

\item \code{one_date}: a date.
}}

