% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yes_or_no.R
\name{yes_or_no}
\alias{yes_or_no}
\title{A binary question to be answered by Yes or No.}
\usage{
yes_or_no(question)
}
\arguments{
\item{question}{The question to be asked.}
}
\value{
\code{TRUE} for yes and \code{FALSE} for No.
}
\description{
A binary question to be answered by Yes or No.
}
