% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_integer.R
\name{fob_integer}
\alias{fob_integer}
\alias{fob_numeric}
\alias{fob_date}
\alias{fob_character}
\alias{fob_yorn}
\alias{fob_boolean}
\title{A binary question to be answered by Yes or No.}
\usage{
fob_integer(question, field_name = "", prompt = NULL)

fob_numeric(question, field_name = "", prompt = NULL)

fob_date(question, field_name = "", format = "\%Y-\%m-\%d", prompt = NULL, ...)

fob_character(question, field_name = "", prompt = NULL)

fob_yorn(question, field_name = "", prompt = "Enter [Y]es or [N]o: ", ...)

fob_boolean(
  question,
  field_name = "",
  prompt = "Enter [T]RUE or [F]ALSE: ",
  ...
)
}
\arguments{
\item{question}{The question to be asked.}

\item{field_name}{Name of the field (optional).}

\item{prompt}{the string printed when prompting the user for input (see \code{\link[=readline]{readline()}}).}

\item{format}{date format.}

\item{...}{further arguments to be passed to \code{\link[=generate_form_pattern]{generate_form_pattern()}}.}
}
\value{
Return the choice selected either as the choice value or its identifier (an integer).
}
\description{
A binary question to be answered by Yes or No.
}
\section{Functions}{
\itemize{
\item \code{fob_numeric}: a numeric.

\item \code{fob_date}: a date.

\item \code{fob_character}: a character.

\item \code{fob_yorn}: \code{TRUE} for yes and \code{FALSE} for no.

\item \code{fob_boolean}: one boolean/logical (i.e. \code{TRUE} or \code{FALSE}).
}}

