% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fob_output.R
\name{fob_output}
\alias{fob_output}
\title{Add an output.}
\usage{
fob_output(FUN, field_name = NULL, ...)
}
\arguments{
\item{FUN}{a function a character string that will be displayed.}

\item{field_name}{field_name.}

\item{...}{arguments passed to \code{FUN}.}
}
\value{
An object of class \code{output}.
}
\description{
Add an output.
}
