% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_from_file.R
\name{form_from_file}
\alias{form_from_file}
\title{Create a form from a connection.}
\usage{
form_from_file(con)
}
\arguments{
\item{con}{a connection object or a character string (see \code{\link[=readLines]{readLines()}}).}
}
\description{
Create a form from a connection.
}
\details{
This function reas
\itemize{
\item \verb{q:} question
\item \verb{t:} type
\item \verb{p:} prompt message (optional)
\item \verb{f:} field name (optional)
\item \verb{#'} a section
}
}
