% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_form}
\alias{generate_form}
\alias{generate_form_pattern}
\alias{generate_form_choices}
\title{Generate form}
\usage{
generate_form(
  prompt = "",
  question = NULL,
  choices = NULL,
  validate = NULL,
  confirm = FALSE,
  pre = NULL,
  post = NULL,
  field_name = ""
)

generate_form_pattern(
  prompt = "",
  question = NULL,
  choices = NULL,
  pattern = "*",
  confirm = FALSE,
  pre = NULL,
  post = NULL,
  field_name = ""
)

generate_form_choices(
  prompt = "",
  question = NULL,
  choices,
  confirm = FALSE,
  pre = NULL,
  post = NULL,
  field_name = ""
)
}
\arguments{
\item{prompt}{The string printed when prompting the user for input (see \code{\link[=readline]{readline()}}).}

\item{question}{The question to be asked.}

\item{choices}{A vector of choices.}

\item{validate}{A function used to validate the answer, if \code{NULL} then validation step is ignored.}

\item{confirm}{a logical. Should the answer be confirmed (default is FALSE).}

\item{pre}{A function for pre processing.}

\item{post}{A function for post processing.}

\item{field_name}{Name of the field (optional).}

\item{pattern}{pattern to be matched.}
}
\description{
Generate form
}
\section{Functions}{
\itemize{
\item \code{generate_form_pattern}: match pattern.

\item \code{generate_form_choices}: list fo choices.
}}

