% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\alias{\%+\%}
\title{objects of class \code{form_partial} or \code{form}}
\usage{
join(...)

x \%+\% y
}
\arguments{
\item{...}{object(s) of class \code{form_partial} or \code{form}.}

\item{x, y}{objects of class \code{form_partial} or \code{form}.}
}
\value{
\code{TRUE} for yes and \code{FALSE} for No.
}
\description{
objects of class \code{form_partial} or \code{form}
}
\section{Functions}{
\itemize{
\item \code{\%+\%}: join operator.
}}

